Rewrite to cogs
Seperated the commands into different categories.
Added command to view all settings
Removed Core
